Reject non-finite samples where they enter, not where they hurt
Only OnsetEnvelope defended against a NaN in the audio, because that is
where PR #54 noticed the damage. The buffer that handed it one did not,
and neither did anything else reading from it.

The gap is not theoretical, and it is worse in the harmony path than in
the onset path the guard was written for. Chroma.estimateTuning
histograms peak deviations and takes the mode; a NaN deviation floors
into slot 0, `best` starts at 0, and `>` against NaN is false, so the
mode can never move off it and the function returns -0.4875 semitones
whatever the recording is. Measured on a 32-second I-V-vi-IV click
track, changing one sample of 705,600: C G Am F at 0.91 confidence
becomes a single N.C. at 0.44, with beat tracking untouched, so it does
not look like a bad sample -- it looks like an unrecognisable recording.

Rejected rather than sanitised, and in the constructor rather than at
the decode boundary, because no decodable file can carry a non-finite
sample: AudioDecoder converts every format to 16-bit signed PCM, and a
float WAV holding NaN, both infinities and 1e30 decodes to zeros. So
this cannot turn a readable recording into a hard error; the only thing
it can catch is a caller with a bug, and substituting zero for that
would alter the audio and hide the cause at once. AudioPipelineTest
asserts the decode behaviour as a tripwire: if a float passthrough is
ever added, the constructor starts throwing there and the exemption
stops being silent.

Two doors that validating only the constructor would have left open,
both measured rather than assumed:

- Spectrogram.compute of a *finite* 1e38 buffer produced 3341
  non-finite bins, so the buffer's invariant does not carry to the
  spectrogram for free. Its magnitudes are now checked too, at 16.6 ms
  against the 531 ms it takes to compute them.
- Resampler.resample turned finite input into 1998 of 2000 infinities
  when upsampling near Float.MAX_VALUE, because it subtracted two
  floats before widening. Downsampling hid it, since the low-pass runs
  first. AudioDecoder resamples before it constructs the buffer, so
  without this the buffer's own check would have reported a fault the
  resampler had just introduced.

Cost is 3.2 ms to scan five minutes of audio, against 531 ms for
Spectrogram.compute and 1124 ms for OnsetEnvelope.fromAudio on the same
samples. There is deliberately no unchecked back door for internally
derived buffers, because a back door is how the invariant stops being
one.

Three of PR #54's tests poisoned *audio* to exercise antiAlias, which
this makes impossible to construct. They are replaced by one asserting
the input is now refused, plus the realistic version of what they
modelled -- a dropout is digital silence, not NaN, and silence still
tracks 120 BPM. antiAlias's run scanning stays pinned by the tests that
call it directly with hand-built mel bands, which is the layer it works
at. The guard itself is left alone: it is unreachable through the
pipeline now, not dead, and removing it is #76. Chroma's own exposure
is #77.

Closes #61

+++ mw-audio/src/main/java/dev/olivelli/musicwizard/audio/Resampler.java
@@ -36,6 +36,15 @@ public final class Resampler {
      * <p>Interpolation is linear. It is not the best kernel available, but the
      * error it leaves is broadband and small relative to the spectral resolution
      * every stage downstream actually uses, whereas aliasing is neither.
+     *
+     * <p>Finite input gives finite output. That is worth stating because it did
+     * not used to be true: the interpolation subtracted two floats before
+     * widening, so upsampling a signal near {@code Float.MAX_VALUE} overflowed
+     * to infinity -- 1998 of 2000 output samples, from input that was entirely
+     * finite. Downsampling hid it, because the low-pass runs first and shrinks
+     * the values. {@link AudioDecoder} resamples before it constructs the
+     * {@link AudioBuffer}, so without this the buffer's own check would have
+     * been the thing reporting a fault the resampler had just introduced.
      */
     public static float[] resample(float[] samples, int fromRate, int toRate) {
         if (fromRate <= 0 || toRate <= 0) {
@@ -65,8 +74,8 @@ public final class Resampler {
             double position = i * step;
             int index = (int) position;
             double fraction = position - index;
-            float a = source[Math.min(index, source.length - 1)];
-            float b = source[Math.min(index + 1, source.length - 1)];
+            double a = source[Math.min(index, source.length - 1)];
+            double b = source[Math.min(index + 1, source.length - 1)];
             out[i] = (float) (a + (b - a) * fraction);
         }
         return out;
